Clean up pcf7931
@iceman1001 s cleanup of pcf7931 commands on client side
and split out pcf7931 from lfops.c as it is getting large
